vendor.nxp: freeze nxpnfc@1.0
It is being used on released devices in P and so must
be frozen. This interface is only used for
vendor<->vendor communication. It's highly recommended
in the future to replace with a non-HIDL interface.

vendor <-> vendor communication is unrestricted. HIDL exists
in its strict form in order to solidify the system<->vendor
boundary. Using it internally within the framework or within
vendor code (where that level of interface isn't required
because it's an implementation detail) will add an undue
burden to maintain the interface. If the interface is already
used system <-> vendor, then of course, the existence of the
HIDL interface is justified and using it system<->system or
vendor<->vendor is fine. Without a pre-existing HIDL interface,
in order to allow code to grow well into the future in the
simplest way possible, it makes sense to put vendor <-> vendor
or system <-> system code into its own abstraction layer
separate from HIDL.

Note that this interface must not be used for system<->vendor
transactions and should be updated to correspond to this:
https://source.android.com/devices/architecture/hidl/versioning#structuring
if it is to be used accross an ABI boundary.

Bug: 77854142
Test: change interface and make sure change is caught
Test: vts_treble_vintf_test
Change-Id: I75ae35654fd7024210406ce37fc607cd225831b4