Fix to issue with NFC VTS-CoreInitiazed test
VTS-CoreInitialized test failed on P17 because PN553/PN81A
was in NCI1.0 mode and setting NCI2.0 configurations
failed.

Fix: If chipType is PN81A(PN553_TC) and if the chip is in
NCI 1.0 mode, Force it to 2.0 mode. To confirm it is
PN553_TC/PN81A chip, FW version check is also added

Bug: 77998208
Bug: 109935766
Test: NFC Enable/Disable, VTS
Change-Id: I78305d33c7263ad1f391af998153f6dade436c7a