Move hidl_package_root to actual path.
In "-r A:B", "A" is the package root, and "B" is the package path.
Together, they tell hidl-gen how hal modules are grouped. Previously,
the path for these was required explicitly. However, now the path can
be inferred automatically. Also, since {B}/current.txt is the location
for API hashes, in order for the build system to properly keep track
of it, we need the hidl package root to contain the current.txt file
(modules can't reference files defined outside of them or their
dependencies). This will also allow deprecation of the now-unneeded
'path' variable.

Bug: 119117395
Test: m nothing (checks this)
Change-Id: Ie2b2fe2b038153aa0799819470ba199da66e45a8